basic table and directory creation
evolved the command structure to parse and interpret the commands  into usable information. This allows for the basic creation functionality along side error catching built in. Most error catching involves not being able to create something that already exits and not being able to create something outside the given directory